lently-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
httpx
firebase-admin
google-api-python-client
google-auth-oauthlib
google-generativeai
slowapi
